feat(twenty-front): give record index pages the settings card layout
Generalizes the settings card chrome into a shared PageCardLayout +
PageCardHeader (with an icon slot) and applies it to record index pages,
so the side panel sits beside a rounded card — consistent with settings —
instead of full-height against a flat table.

- New PageCardLayout (rounded card: primary bar, optional secondary bar,
  body) and PageCardHeader (centered title with an icon slot + optional
  breadcrumb) under @/ui/layout/page. SettingsPageLayout now delegates to
  them; the settings-only SettingsPageHeader is removed.
- Record index: the page header (object icon + label + actions) becomes the
  card primary bar; the view bar (view picker + filter/sort/options) is
  lifted into the secondary bar (RecordIndexViewBar); the table/board/
  calendar is the card body.
